+++ ansible_mitogen/transport_config.py
@@ -79,6 +79,11 @@ import mitogen.core
 
 LOG = logging.getLogger(__name__)
 
+if ansible_mitogen.utils.ansible_version[:2] >= (2, 21):
+    _INTERPRETER_DISCOVERY_MODES = frozenset(['auto', 'auto_silent'])
+else:
+    _INTERPRETER_DISCOVERY_MODES = frozenset(['auto', 'auto_legacy', 'auto_silent', 'auto_legacy_silent'])
+
 if ansible_mitogen.utils.ansible_version[:2] >= (2, 19):
     _FALLBACK_INTERPRETER = ansible.executor.interpreter_discovery._FALLBACK_INTERPRETER
 elif ansible_mitogen.utils.ansible_version[:2] >= (2, 17):
@@ -98,7 +103,7 @@ def run_interpreter_discovery_if_necessary(s, candidates, task_vars, action, red
     if action._mitogen_discovering_interpreter:
         return action._mitogen_interpreter_candidate
 
-    if s in ['auto', 'auto_legacy', 'auto_silent', 'auto_legacy_silent']:
+    if s in _INTERPRETER_DISCOVERY_MODES:
         # python is the only supported interpreter_name as of Ansible 2.8.8
         interpreter_name = 'python'
         discovered_interpreter_config = u'discovered_interpreter_%s' % interpreter_name